fastapi/requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
h11==0.12.0
html2text==2020.1.16
idna==3.2
Markdown==3.3.4
markdownify==0.9.4
mysql-connector-python==8.0.26
protobuf==3.17.3
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
python-dateutil==2.8.2
python-decouple==3.4
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
